Polish

+++ webauthn4j-ctap/src/main/kotlin/com/webauthn4j/ctap/core/data/ble/BLEInitializationFrameFragment.kt
@@ -4,7 +4,7 @@ import com.webauthn4j.util.ArrayUtil
 import com.webauthn4j.util.UnsignedNumberUtil
 import java.nio.ByteBuffer
 
-class BLEInitializationFrameFragment(val cmd: BLEFrameCommand, val length: Int, data: ByteArray) :
+class BLEInitializationFrameFragment(val cmd: BLEFrameCommand, val length: UShort, data: ByteArray) :
     BLEFrameFragment {
 
     override val data: ByteArray
@@ -14,7 +14,7 @@ class BLEInitializationFrameFragment(val cmd: BLEFrameCommand, val length: Int, 
         get() {
             val byteBuffer = ByteBuffer.allocate(3 + data.size)
             byteBuffer.put(cmd.value)
-            byteBuffer.put(UnsignedNumberUtil.toBytes(length))
+            byteBuffer.put(UnsignedNumberUtil.toBytes(length.toInt()))
             byteBuffer.put(data)
             return byteBuffer.array()
         }
@@ -24,7 +24,7 @@ class BLEInitializationFrameFragment(val cmd: BLEFrameCommand, val length: Int, 
         fun parse(bytes: ByteArray): BLEInitializationFrameFragment {
             val byteBuffer = ByteBuffer.wrap(bytes)
             val cmd = BLEFrameCommand(byteBuffer.get())
-            val length = UnsignedNumberUtil.getUnsignedShort(byteBuffer)
+            val length = UnsignedNumberUtil.getUnsignedShort(byteBuffer).toUShort()
             val data = ByteArray(byteBuffer.remaining())
             byteBuffer[data]
             return BLEInitializationFrameFragment(cmd, length, data)
